+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Maintain bookmarks when editing connection details
 ### Changed
 - Updated dependencies
 - Upgraded build & test environments from nodejs 16 to nodejs 20